+++ src/pages/catalog-product/api/index.ts
@@ -1,3 +1,3 @@
-export { sendingFilterRequest } from "./filter-request";
+export { sendingFilterSortingRequest } from "./filter-sorting-request";
 export { getAllProducts } from "./request-to-get-all-products";
 export * from "./types";
